fix(nc-gui): separate out permissions content from modal file
